FIX keep at least one admin

+++ src/Controller/AdminUserController.php
@@ -71,27 +71,31 @@ class AdminUserController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $userRepository->save($user, true);
-
             $roles = $user->getRoles();
-            if (in_array("ROLE_RECRUITER", $roles)) {
-                if ($user->getRecruiter() === null) {
-                    $recruiter = new Recruiter();
-                    $recruiter->setUser($user);
 
-                    $recruiterRepository->save($recruiter, true);
-                    $this->addFlash("success", "Le rôle recruteur a bien été ajouté");
+            $isUpdatable = $this->isUpdatable($userRepository, $user);
+
+            if ($isUpdatable) {
+                $userRepository->save($user, true);
+
+                if (in_array("ROLE_RECRUITER", $roles)) {
+                    if ($user->getRecruiter() === null) {
+                        $recruiter = new Recruiter();
+                        $recruiter->setUser($user);
+
+                        $recruiterRepository->save($recruiter, true);
+                    }
                 }
-            }
 
-            if (in_array("ROLE_CANDIDATE", $roles)) {
-                if ($user->getCandidate() === null) {
-                    $candidate = new Candidate();
-                    $candidate->setUser($user);
+                if (in_array("ROLE_CANDIDATE", $roles)) {
+                    if ($user->getCandidate() === null) {
+                        $candidate = new Candidate();
+                        $candidate->setUser($user);
 
-                    $candidateRepository->save($candidate, true);
-                    $this->addFlash("success", "Le rôle candidat a bien été ajouté");
+                        $candidateRepository->save($candidate, true);
+                    }
                 }
+                $this->addFlash("success", "Le rôle a bien été mis à jour");
             }
 
             return $this->redirectToRoute('app_admin_index', [], Response::HTTP_SEE_OTHER);
@@ -106,12 +110,35 @@ class AdminUserController extends AbstractController
     #[Route('/{id}', name: 'app_admin_delete', methods: ['POST'])]
     public function delete(Request $request, User $user, UserRepository $userRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
-            $userRepository->remove($user, true);
+        $admins = $userRepository->findAllAdmin();
 
-            $this->addFlash("success", "Utilisateur supprimé avec succès.");
+        if (count($admins) > 1) {
+            if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
+                $userRepository->remove($user, true);
+
+                $this->addFlash("success", "Utilisateur supprimé avec succès.");
+            }
+        } else {
+            $this->addFlash("danger", "Impossible de supprimer le profil administrateur");
         }
 
         return $this->redirectToRoute('app_admin_index', [], Response::HTTP_SEE_OTHER);
+    }
+
+
+    private function isUpdatable(UserRepository $userRepository, User $user,): bool
+    {
+        $admins = $userRepository->findAllAdmin();
+
+        if (count($admins) === 1) {
+            if ($admins[0]->getId() === $user->getId()) {
+                if (!in_array("ROLE_ADMIN", $admins[0]->getRoles())) {
+                    $this->addFlash("danger", "Impossible de supprimer le profil administrateur");
+                    return false;
+                }
+            }
+        }
+
+        return true;
     }
 }
